resource logic moved to Trait
